orthos/backend/requirements.txt
flask==2.0.1
firebase-admin==5.0.3
requests==2.26.0
python-dotenv==0.19.1
numpy==1.21.2
flask-cors==3.0.10